refactor: app.service.SocialTokenService -> domain.provider.SocialTokenProvider 인터페이스 이름 변경 및 패키지 이동
